requirements.txt
aiohttp==3.1.0
async-timeout==2.0.1
attrs==17.4.0
chardet==3.0.4
idna==2.6
idna-ssl==1.0.1
multidict==4.1.0
psutil==5.4.3
python-dateutil==2.7.1
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
uptime==3.0.1
websockets==4.0.1
yarl==1.1.1
